requirements.txt
feedparser
requests
nest_asyncio
python-telegram-bot==20.7
